+++ server/utils/responses/__init__.py
@@ -1,0 +1,1 @@
+HTTP_404_DETAIL = "Resource not found"
