+++ server/utils/responses/errors.py
@@ -1,0 +1,55 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+from responses import HTTP_404_DETAIL
+
+
+### Error responses
+class ErrorResponse(BaseModel, frozen=True):
+    result: str = "error"
+    detail: str
+
+
+### Responses for HTTP 400-499 status codes
+
+
+# HTTP 400
+class BadRequestResponse(ErrorResponse, frozen=True):
+    pass
+
+
+# HTTP 401
+class UnauthorizedResponse(ErrorResponse, frozen=True):
+    pass
+
+
+# HTTP 403
+class ForbiddenResponse(ErrorResponse, frozen=True):
+    pass
+
+
+# HTTP 404
+class NotFoundResponse(ErrorResponse, frozen=True):
+    detail = HTTP_404_DETAIL
+
+
+# HTTP 409
+class ConflictResponse(ErrorResponse, frozen=True):
+    pass
+
+
+# HTTP 400/422
+class RequestValidationErrorDetails(BaseModel, frozen=True):
+    detail: str
+    context: Optional[str]
+
+
+class RequestValidationErrorResponse(BaseModel, frozen=True):
+    result: str = "error"
+    errors: list[RequestValidationErrorDetails]
+
+
+# Any status codes
+class HTTPExceptionResponse(ErrorResponse, frozen=True):
+    pass
